# openapi.yaml
openapi: 3.0.3
info:
  title: Hoobox Orders API
  description: |
    API de pedidos com processamento assíncrono (BullMQ).

    - `POST /orders` persiste o pedido como `PENDING` e enfileira `order.created`.
    - O worker processa em ~1–2s (estoque / regras de falha).
    - `customerName` contendo `fail` (case insensitive) simula falha com retry até `FAILED`.
  version: 1.0.0

servers:
  - url: http://localhost:3333
    description: Desenvolvimento local

tags:
  - name: health
  - name: orders

paths:
  /:
    get:
      tags: [health]
      summary: Health / hello
      operationId: getHello
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: Hello World!

  /orders:
    get:
      tags: [orders]
      summary: Listar pedidos (paginado)
      operationId: listOrders
      parameters:
        - name: page
          in: query
          description: Página (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Itens por página (default 10, máx. 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Lista paginada (com itens de cada pedido)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
        "400":
          $ref: "#/components/responses/ValidationError"

    post:
      tags: [orders]
      summary: Criar pedido
      description: |
        Calcula total, salva pedido `PENDING`, publica job na fila e retorna sucesso imediato.
        Produtos devem existir no banco (campo `name`).
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            examples:
              ok:
                summary: Pedido válido
                value:
                  customerName: Maria Silva
                  items:
                    - productName: Camiseta
                      quantity: 2
                      price: 19.99
              forcedFail:
                summary: Falha simulada (retry → FAILED)
                value:
                  customerName: Cliente fail test
                  items:
                    - productName: Camiseta
                      quantity: 1
                      price: 19.99
      responses:
        "201":
          description: Pedido aceito para processamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"

  /orders/{id}:
    get:
      tags: [orders]
      summary: Consultar pedido por ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Detalhe do pedido (com itens)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Pedido não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpError"

components:
  schemas:
    OrderStatus:
      type: string
      enum: [PENDING, PROCESSED, FAILED]

    CreateOrderItem:
      type: object
      required: [productName, quantity, price]
      properties:
        productName:
          type: string
          minLength: 1
          example: Camiseta
        quantity:
          type: integer
          minimum: 1
          example: 2
        price:
          type: number
          format: double
          minimum: 0.01
          maximum: 21474836
          description: Preço unitário em reais (até 2 casas decimais)
          example: 19.99

    CreateOrderRequest:
      type: object
      required: [customerName, items]
      properties:
        customerName:
          type: string
          minLength: 1
          example: Maria Silva
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateOrderItem"

    CreateOrderResponse:
      type: object
      properties:
        status:
          type: string
          example: ok

    OrderItem:
      type: object
      properties:
        productName:
          type: string
        quantity:
          type: integer
        priceCents:
          type: integer
          description: Preço unitário em centavos
        price:
          type: string
          description: Preço formatado (ex. "19.99")
          example: "19.99"

    OrderSummary:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        totalCents:
          type: integer
        total:
          type: string
          example: "39.98"
        status:
          $ref: "#/components/schemas/OrderStatus"
        failureReason:
          type: string
          nullable: true
          example: estoque insuficiente
        createdAt:
          type: string
          format: date-time

    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/OrderSummary"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"

    PaginatedOrders:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrderDetail"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    HttpError:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string

  responses:
    ValidationError:
      description: Erro de validação (DTO)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HttpError"
          example:
            statusCode: 400
            message: ["price must be a positive number"]
            error: Bad Request